Changed name of customerProductCatalog
