+++ HipAndClavicle/HipAndClavicle.csproj
@@ -60,7 +60,7 @@
 		<None Remove="Views\Contact\" />
 	</ItemGroup>
 	<ItemGroup>
-		<None Include="Views\CustomerProductCatalog\SearchByColor.cshtml" />
+		<None Include="Views\Customer\SearchByColor.cshtml" />
 		<None Include="wwwroot\css\Product.css" />
 	</ItemGroup>
 	<ItemGroup>
