--- HipAndClavicle/Views/CustomerProductCatalog/SearchByColor.cshtml
@@ -1,41 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Shop by Color Family";
-}
-
-<h1>Shop by color family</h1>
-
-<div class="container">
-    <div class="row">
-        <a asp-action="CustFindListings" asp-route-colorFamilyName="Blues" class="col col-md-6 col-sm-12 text-decoration-none text-dark mb-5">
-            <img src="../images/blue01.png" class="img-fluid"/>
-            <h3 class="">Blue</h3>
-        </a>
-
-        <a asp-action="CustFindListings" asp-route-colorFamilyName="Reds" class="col col-md-6 col-sm-12 text-decoration-none text-dark mb-5">
-            <img src="../images/red01.png" class="img-fluid"/>
-            <h3>Red</h3>
-        </a>
-
-        <a asp-action="CustFindListings" asp-route-colorFamilyName="Yellows" class="col col-md-6 col-sm-12 text-decoration-none text-dark mb-5">
-            <img src="../images/yellow01.png" class="img-fluid" />
-            <h3>Yellow</h3>
-        </a>
-
-        <a asp-action="CustFindListings" asp-route-colorFamilyName="Greens" class="col col-md-6 col-sm-12 text-decoration-none text-dark mb-5">
-            <img src="../images/green01.png" class="img-fluid" />
-            <h3>Green</h3>
-        </a>
-
-        <a asp-action="CustFindListings" asp-route-colorFamilyName="Oranges" class="col col-md-6 col-sm-12 text-decoration-none text-dark mb-5">
-            <img src="../images/orange02.png" class="img-fluid" />
-            <h3>Orange</h3>
-        </a>
-
-        <a asp-action="CustFindListings" asp-route-colorFamilyName="Purples" class="col col-md-6 col-sm-12 text-decoration-none text-dark mb-5">
-            <img src="../images/purple01.png" class="img-fluid" />
-            <h3>Purple</h3>
-        </a>
-    </div>
-
-
-</div>
