+++ HipAndClavicle/Views/Shared/_Layout.cshtml
@@ -27,11 +27,11 @@
                         </li>
 
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="CustomerProductCatalog" asp-action="SearchByColor">Shop by Color</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Customer" asp-action="SearchByColor">Shop by Color</a>
                         </li>
 
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="CustomerProductCatalog" asp-action="CustFindListings">Shop All Products</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Customer" asp-action="CustFindListings">Shop All Products</a>
                         </li>
 
                         <li class="nav-item">
